requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
pgvector>=0.2.0
google-generativeai>=0.8.0
numpy>=1.26.0
httpx>=0.25.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
